+++ packages/@zenithui/utils/src/index.ts
@@ -9,6 +9,10 @@ export { default as useMediaQuery } from "./hooks/use-media-query"
 export { default as useTheme } from "./hooks/use-theme"
 export { default as useUnMount } from "./hooks/use-unmount"
 export {
+  default as useThemeMode,
+  type ThemeAnimationType,
+} from "./hooks/use-theme-mode"
+export {
   useDebounce,
   usePrevious,
   useDefault,