+++ packages/@zenithui/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zenithui/utils
 
+## 0.3.0
+
+### Minor Changes
+
+- Addded theme toggle hook
+
 ## 0.2.3
 
 ### Patch Changes